+++ packages/genesis/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @edenapp/genesis
 
+## 0.8.0
+
+### Minor Changes
+
+- Release 0.8.0.
+
+### Patch Changes
+
+- Updated dependencies
+  - @edenapp/types@0.8.0
+
 ## 0.7.3
 
 ### Patch Changes