+++ packages/sdk/apps/com/eden/process-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # com.eden.process-manager
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @edenapp/babel@0.8.0
+
 ## 1.0.8
 
 ### Patch Changes